refactored and fixed attachment gen
